requirements.txt
requests
python3-linkedin
pytest
backoff
asyncio
ttkbootstrap